move transporters to their own files
